Changed url to folder/file
By default the theme builds pages to the root folder. Now, it builds to a folder for each subject.

Filepaths needed to be changed. An important filepath is that Liquid matches the url of the page with the url of the sidebar item, so those need to be the same and so you have to go one folder back when moving between pages.